auth update

+++ src/chirp/chirp.controller.ts
@@ -1,14 +1,29 @@
-import { Controller, Delete, Get, Param, Put, Req } from '@nestjs/common';
+import {
+  Body,
+  Controller,
+  Delete,
+  Get,
+  Param,
+  Put,
+  Req,
+  UnauthorizedException,
+  UseGuards,
+} from '@nestjs/common';
 import { ChirpService } from './chirp.service';
 import { ChirpDto } from './dto/chirp.dto';
+import { LocalGuard } from '../auth/local.guard';
 
 @Controller('chirp')
 export class ChirpController {
   constructor(private chirpService: ChirpService) {}
 
   @Put()
-  create(chirpDto: ChirpDto, @Req() req: Express.Request) {
-    chirpDto.authorId = req.user['id'];
+  @UseGuards(LocalGuard)
+  create(@Body() reqBody: any, @Req() req: Express.Request) {
+    const chirpDto: ChirpDto = {
+      authorId: req.user['id'],
+      content: reqBody.content,
+    };
     return this.chirpService.create(chirpDto);
   }
 
@@ -28,7 +43,14 @@ export class ChirpController {
   }
 
   @Delete(':id')
-  delete(@Param('id') id: string) {
-    return this.chirpService.delete(+id);
+  @UseGuards(LocalGuard)
+  async delete(@Param('id') id: string, @Req() req: Express.Request) {
+    if (
+      (await this.chirpService.findById(+id)).author['id'] === req.user['id']
+    ) {
+      return this.chirpService.delete(+id);
+    } else {
+      throw new UnauthorizedException();
+    }
   }
 }
